// src/interfaces/Student.tsx
export interface Student {
    id: number,
    first_name: string;
    last_name: string;
    email: string;
    phone: string;
    address_line_1: string;
    address_line_2: string;
    address_city: string;
    address_state: string;
    address_zip: string;
    password: string;
    username: string;
}